backend/requirements.txt
# Sentinel Authority Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
starlette==0.35.1

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client (for external integrations)
httpx==0.26.0
aiohttp==3.9.1

# Caching & Rate Limiting
redis==5.0.1
aioredis==2.0.1

# Background Tasks
celery==5.3.4

# Storage
boto3==1.34.14
aioboto3==12.3.0

# Monitoring & Logging
structlog==24.1.0
prometheus-client==0.19.0

# Email
aiosmtplib==3.0.1

# Billing
stripe==7.10.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
